refactor(repo/config) move config under repo
