+++ commands/http/client.go
@@ -11,7 +11,7 @@ import (
 	"strings"
 
 	cmds "github.com/jbenet/go-ipfs/commands"
-	config "github.com/jbenet/go-ipfs/config"
+	config "github.com/jbenet/go-ipfs/repo/config"
 )
 
 const (
