+++ core/commands/mount_unix.go
@@ -9,7 +9,7 @@ import (
 	"time"
 
 	cmds "github.com/jbenet/go-ipfs/commands"
-	config "github.com/jbenet/go-ipfs/config"
+	config "github.com/jbenet/go-ipfs/repo/config"
 	core "github.com/jbenet/go-ipfs/core"
 	ipns "github.com/jbenet/go-ipfs/fuse/ipns"
 	mount "github.com/jbenet/go-ipfs/fuse/mount"
